+++ packages/authentication/modules/user/getUser/__tests__/getUser.test.ts
@@ -8,7 +8,7 @@ import jwt from './fixtures/jwt.json'
 describe('getUser', () => {
   it('should return the user from the JWT cookie', async () => {
     ;(Cookies.get as CookiesGetByNameFn) = jest.fn(() => jwt.token)
-    const user = await getUser()
+    const user = getUser()
 
     expect(user?.email).toBe('user@company.com')
     expect(user?.firstName).toBe('John')
@@ -17,7 +17,7 @@ describe('getUser', () => {
 
   it('should return null if the JWT cookie is not set', async () => {
     ;(Cookies.get as CookiesGetByNameFn) = jest.fn(() => undefined)
-    const user = await getUser()
+    const user = getUser()
 
     expect(user).toBeNull()
   })