+++ packages/authentication/modules/user/getUser/types.ts
@@ -1,3 +1,5 @@
+import { ServerSideRenderingOption } from '@baseapp-frontend/utils'
+
 import { ICookieName } from '../../../types/auth'
 
-export interface GetUserOptions extends ICookieName {}
+export interface GetUserOptions extends ICookieName, ServerSideRenderingOption {}