+++ packages/utils/functions/token/getToken/index.ts
@@ -1,11 +1,8 @@
-import ClientCookies from 'js-cookie'
-
 import { ACCESS_COOKIE_NAME } from '../../../constants/cookie'
+import type { ServerSideRenderingOption } from '../../../types/server'
+import { getCookie } from '../../cookie'
 
-export const getToken = async (cookieName = ACCESS_COOKIE_NAME) => {
-  if (typeof window === typeof undefined) {
-    const { cookies: serverCookies } = await import('next/headers')
-    return serverCookies().get(cookieName)?.value
-  }
-  return ClientCookies.get(cookieName)
-}
+export const getToken = (
+  cookieName = ACCESS_COOKIE_NAME,
+  { noSSR = false }: ServerSideRenderingOption = {},
+) => getCookie<string>(cookieName, { noSSR })